refactor: move schema tests to dedicated files
